Refactored code and tests
